Created policies and implemented them in the controllers

+++ app/Http/Controllers/EntryController.php
@@ -24,7 +24,8 @@ class EntryController extends Controller
      */
     public function create(Project $project, Task $task)
     {
-        abort_unless($task->project_id === $project->id, 404);
+        // abort_unless($task->project_id === $project->id, 403); redundant code
+        $this->authorize('create', $project);
         return view('entries.create', [
         'project' => $project,    
         'task' => $task,
@@ -36,7 +37,7 @@ class EntryController extends Controller
      */
     public function store(Project $project, Task $task, Request $request)
     {
-    
+         $this->authorize('create', $project);
         $validated = $request->validate([
             'work_date' => ['required', 'date'],
             'minutes' => ['required', 'integer', 'min:1', 'max:1440'],
@@ -73,6 +74,7 @@ class EntryController extends Controller
      */
     public function edit(Project $project, Task $task, Entry $entry)
     {
+         $this->authorize('update', $project);
         return view('entries.edit', compact( 'project','task','entry'));
     }
 
@@ -81,6 +83,7 @@ class EntryController extends Controller
      */
     public function update(Request $request, Project $project, Task $task, Entry $entry)
     {
+        $this->authorize('update', $project);
         $validated = $request->validate([
             'work_date' => ['required', 'date'],
             'minutes' => ['required', 'integer', 'min:1', 'max:1440'],
@@ -96,11 +99,7 @@ class EntryController extends Controller
      */
     public function destroy(Project $project, Task $task, Entry $entry)
     {
-        abort_unless(
-    $project->id === $task->project_id && $task->id === $entry->task_id,
-    404
-        );
-
+        $this->authorize('delete', $project);
         $entry->delete();
         return back();
     }
